move helpers.GetTermWinsize -> osutil.GetTermWinsize
