+++ osutil/winsize.go
@@ -17,7 +17,7 @@
  *
  */
 
-package helpers
+package osutil
 
 import (
 	"syscall"
